Move TimeZoneSetter middleware in proper place
